# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/preferences": {
            "get": {
                "description": "Get a list of all notification preferences",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notification_preferences"
                ],
                "summary": "Get all notification preferences",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entities.NotificationPreferencesResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new notification preference with the given details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notification_preferences"
                ],
                "summary": "Create a new notification preference",
                "parameters": [
                    {
                        "description": "Preference request body",
                        "name": "preference",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entities.NotificationPreferencesCrRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/entities.NotificationPreferencesResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/preferences/{id}": {
            "get": {
                "description": "Get details of a specific notification preference by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notification_preferences"
                ],
                "summary": "Get notification preference by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Preference ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.NotificationPreferencesResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Update details of an existing notification preference by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notification_preferences"
                ],
                "summary": "Update a notification preference",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Preference ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Preference request body",
                        "name": "preference",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entities.NotificationPreferencesUpRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.NotificationPreferencesResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a notification preference by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notification_preferences"
                ],
                "summary": "Delete a notification preference",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Preference ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entities.NotificationPreferencesCrRequest": {
            "type": "object",
            "required": [
                "receive_emails",
                "user_id"
            ],
            "properties": {
                "receive_emails": {
                    "type": "boolean"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "entities.NotificationPreferencesResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "receive_emails": {
                    "type": "boolean"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "entities.NotificationPreferencesUpRequest": {
            "type": "object",
            "properties": {
                "receive_emails": {
                    "type": "boolean"
                }
            }
        },
...
    }
}